REQ5-Endpoint para crear nuevo evento
Endpoint that allows to create a new event
