+++ Event.js
@@ -1,0 +1,10 @@
+class Event {
+	constructor(id, title, description, date) {
+		this.id = id;
+		this.title=title;
+		this.description=description;
+		this.date=date;
+	}
+}
+
+exports.Event = Event;
