+++ app.js
@@ -2,20 +2,16 @@ const express = require('express')
 const app = express();
 const eventsArray = new Array();
 
-class Event {
-	constructor(id, title, description, date) {
-		this.id = id;
-		this.title=title;
-		this.description=description;
-		this.date=date;
-	}
-}
+const Event = require('./Event').Event;
+const bodyParser = require('body-parser');
 
 eventsArray.push(new Event(1, 'Conferencia', 'Node.js, Pair Programming', '2017-03-15'));
 eventsArray.push(new Event(2, 'Concierto', 'Sinfonica', '2017-03-11'));
 eventsArray.push(new Event(3, 'Cita Dentista', 'Limpieza Dental', '2017-03-09'));
 eventsArray.push(new Event(4, 'Clases de frances', 'Iniciando', '2017-03-01'));
 eventsArray.push(new Event(5, 'Cena', 'Conunidad Agile', '2017-02-15'));
+
+app.use(bodyParser.json());
 
 app.get('/', (req, res) => {
   res.send('Got a GET request!')
@@ -33,7 +29,7 @@ app.get('/events/:id', (req, res) => {
 	let idBuscado = req.params.id;
 	 try
 	 {
-		let eventoEncontrado = eventsArray.find(o => o.id == idBuscado);
+		let eventoEncontrado = eventsArray.find(evento => evento.id == idBuscado);
 
 		if (eventoEncontrado == undefined)  {
 			res.status(404).send('El evento no existe.');
@@ -43,6 +39,23 @@ app.get('/events/:id', (req, res) => {
 		}
 	}
 	catch(ex) {
+		res.status(500).send('Se encontró un error.')
+	}
+})
+
+app.post('/events', (req,res) => {
+	try{
+
+		let idEncontrado = eventsArray.find(evento => evento.id == req.body.id);
+
+		if(idEncontrado == undefined){
+			let newEvent = new Event(req.body.id, req.body.title, req.body.description, req.body.date);
+			eventsArray.push(newEvent);
+			res.status(201).json(newEvent);
+		}else{
+			res.status(409).send('Ya existe un evento con ese id');
+		}
+	}catch(ex){
 		res.status(500).send('Se encontró un error.')
 	}
 })
