+++ package.json
@@ -17,6 +17,7 @@
   },
   "homepage": "https://github.com/ccalvarez/EmventsInc",
   "dependencies": {
+    "body-parser": "^1.17.1",
     "express": "4.15.2"
   }
 }
